Dot color node attribute added Label with spaces and new line handled New error message in case of invalid attributes
