+++ doted/major_1/minor_0/observer/Subject.py
