+++ doted/major_1/minor_0/view/widget/MainWindow.py
