+++ doted/major_1/minor_0/view/widget/PropertiesView.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+
+from PyQt5.QtGui import QStandardItemModel
+from PyQt5.QtWidgets import QTreeView
+
+from major_1.minor_0.view.widget.View import View
+
+
+class PropertiesView(View, QTreeView):
+    '''The PropertiesView class defines a view for the properties of a Node.'''
+
+
+    def __init__(self):
+        # Parent constructor(s)
+        View.__init__(self)
+        QTreeView.__init__(self)
+        
+        model = QStandardItemModel()
+        model.setHorizontalHeaderLabels(["Properties", "Values"])
+        self.setModel(model)
