+++ doted/major_1/minor_1/enumeration/NodeDotAttrs.py
@@ -1,0 +1,11 @@
+from enum import Enum
+
+
+class NodeDotAttrs(Enum):
+    '''The NodeDotAttrs enum defines the dot attributes of a Node.
+    Values must take the real name of dot attributes.'''
+    
+    
+    label = "label"
+    pos = "pos"
+    color = "color"
